+++ src/jtl/Connector/Shopware/Controller/CustomerOrder.php
@@ -80,6 +80,11 @@ class CustomerOrder extends DataController
                         $orderItem = Mmc::getModel('CustomerOrderItem');
                         $orderItem->map(true, DataConverter::toObject($detailSW, true));
 
+                        // Tax Free
+                        if ((int) $orderSW['taxFree'] == 1) {
+                            $orderItem->setVat(0.0);
+                        }
+
                         $detail = $productMapper->findDetailBy(array('number' => $detailSW['articleNumber']));
                         if ($detail !== null) {
                             //throw new \Exception(sprintf('Cannot find detail with number (%s)', $detailSW['articleNumber']));
@@ -143,7 +148,7 @@ class CustomerOrder extends DataController
                     // Attributes
                     for ($i = 1; $i <= 6; $i++) {
                         if (isset($orderSW['attribute']["attribute{$i}"]) && strlen($orderSW['attribute']["attribute{$i}"]) > 0) {
-                            $customerOrderAttr = Mmc::getModel('CustomerOrderAttrs');
+                            $customerOrderAttr = Mmc::getModel('CustomerOrderAttr');
                             $customerOrderAttr->map(true, DataConverter::toObject($orderSW['attribute']));
                             $customerOrderAttr->setKey("attribute{$i}")
                                 ->setValue($orderSW['attribute']["attribute{$i}"]);